scripts/scraper/python/requirements.txt
playwright==1.51.0
python-dotenv==1.0.1
supabase==2.13.0
